requirements.txt
Flask==2.3.2
python-telegram-bot==13.17
gunicorn==20.1.0